pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "switchbot-exporter"
version = "1.0.0"
description = "Prometheus SwitchBot Exporter"
authors = [{ name = "mizu", email = "mizu.copo@gmail.com" }]
requires-python = ">=3.13,<4.0"
dependencies = [
    "click~=8.1",
    "requests~=2.32",
    "inflect~=7.4",
    "inflection~=0.5",
    "flask~=3.0",
    "gunicorn~=23.0",
    "python-decouple~=3.8",
]

[project.optional-dependencies]
dev = [
    "black~=24.10",
    "mypy~=1.13",
    "ruff~=0.7",
    "types-requests~=2.32",
    "pytest~=8.3",
]

[tool.black]
line-length = 88
target-version = ["py313"]
skip-string-normalization = false
include = '\.pyi?$'

[tool.ruff]
fix = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "C90",
    "N",
    "I",
    "D",
    "ANN",
]
ignore = [
  "D203",
  "D212",
  "E501",
]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
